Fix compilation errors in AITestRunner and AIController
- Cast Average() to float in AITestRunner.cs
- Change Card.EffectType.None to NoEffect in AIController.cs
- Remove abilityTarget check (field doesn't exist on Card)

+++ TarotBattlegrounds-POC/Assets/Scripts/AI/AITestRunner.cs
@@ -99,7 +99,7 @@ public class AITestRunner : MonoBehaviour
             yield return new WaitForSeconds(delayBetweenGames);
         }
 
-        avgGameLength = gameLengths.Count > 0 ? gameLengths.Average() : 0f;
+        avgGameLength = gameLengths.Count > 0 ? (float)gameLengths.Average() : 0f;
         isRunning = false;
         PrintResults();
     }
